Clean the code and separate the pages and also change the the components name and pages name
