+++ src/App.js
@@ -1,12 +1,12 @@
-import logo from './logo.svg';
-import './App.css';
-import Navbar from './components/Navbar';
-import Routes from './Routes';
+import logo from "./logo.svg";
+import "./App.css";
+import { Navbar } from "./components";
+import Routes from "./Routes";
 function App() {
   return (
     <div className="App">
-      <Navbar/>
-      <Routes/>
+      <Navbar />
+      <Routes />
     </div>
   );
 }
