+++ src/Routes.js
@@ -1,14 +1,11 @@
-import React from "react";
-import { Box } from "@mui/system";
-import Container from "@mui/material/Container";
-import Navbar from "./components/Navbar";
-import { Routes as BrowserRoutes, Route, } from "react-router-dom";
-import MoviesMainPage from "./components/MoviesMainPage";
-import MoviesDetailPage from "./components/MoviesDetailPage";
+import { Routes as BrowserRoutes, Route } from "react-router-dom";
+// import { Home, MoviesDetailPage } from "./pages";
+import Home from "./pages/Home";
+import MoviesDetailPage from "./pages/MoviesDetailPage";
 export default function Routes() {
   return (
     <BrowserRoutes>
-      <Route path="/" element={<MoviesMainPage />} />
+      <Route path="/" element={<Home />} />
       <Route path="detail/:id" element={<MoviesDetailPage />} />
     </BrowserRoutes>
   );
