+++ src/components/Header.jsx
@@ -1,11 +1,15 @@
-import React from "react";
 import { useNavigate } from "react-router";
 import { useGetSerchMoviesByNameQuery } from "../service/services";
-import { Grid, TextField, Autocomplete, Typography } from "@mui/material";
+import {
+  Grid,
+  TextField,
+  Autocomplete,
+  Typography,
+  CircularProgress,
+} from "@mui/material";
 import { useState } from "react";
-import CircularProgress from "@mui/material/CircularProgress";
 
-export default function MoviesHeader() {
+export const Header = () => {
   const navigation = useNavigate();
   const [input, setInput] = useState("a");
   const { data, error, isLoading } = useGetSerchMoviesByNameQuery(input);
@@ -92,4 +96,4 @@ export default function MoviesHeader() {
       </Grid>
     </Grid>
   );
-}
+};
